Rework model to include schema property
