+++ controllers/apply-next/simple/form-model.js
@@ -1,11 +1,10 @@
 'use strict';
+const { includes } = require('lodash');
 const path = require('path');
-const { includes, values } = require('lodash');
-const moment = require('moment');
 
-const schema = require('./schema');
+const { schema, allFields, addressFields, organisationTypes } = require('./fields');
 const processor = require('./processor');
-const validators = require('../validators');
+const validatorsLegacy = require('../validators-legacy');
 
 /**
  * @typedef {Object} LocaleString
@@ -42,627 +41,11 @@ const validators = require('../validators');
  * @property {LocaleString} [explanation]
  * @property {String} type
  * @property {Object} [attributes]
- * @property {boolean} [isRequired]
- * @property {function} validator
+ * @property {Boolean} [isRequired]
+ * @property {Function} schema
  */
 
 const SESSION_KEY = 'awards-for-all';
-
-const MIN_APPLICANT_AGE = 16;
-
-const ORGANISATION_TYPES = {
-    constitutedVoluntaryCommunity: {
-        value: 'constituted-voluntary-community',
-        label: { en: 'Constituted voluntary or community organisation', cy: '' },
-        explanation: { en: 'i.e. not registered as a company or charity', cy: '' }
-    },
-    unincorporatedRegisteredCharity: {
-        value: 'unincorporated-registered-charity',
-        label: { en: 'Unincorporated registered charity', cy: '' },
-        explanation: { en: 'You will need to provide your charities commission registration number', cy: '' }
-    },
-    charitableIncorporatedOrganisation: {
-        value: 'charitable-incorporated-organisation',
-        label: { en: 'Charitable incorporated organisation (CIO)', cy: '' },
-        explanation: { en: 'You will need to provide your companies house registration number', cy: '' }
-    },
-    notForProfitCompany: {
-        value: 'not-for-profit-company',
-        label: { en: 'Not-for-profit company', cy: '' }
-    },
-    communityInterestCompany: {
-        value: 'community-interest-company',
-        label: { en: 'Community interest company (CIC)', cy: '' }
-    },
-    school: {
-        value: 'school',
-        label: { en: 'School', cy: '' }
-    },
-    statutoryBody: {
-        value: 'statutory-body',
-        label: { en: 'Statutory body', cy: '' }
-    }
-};
-
-function postcodeField(props) {
-    // Allows us to use postcode validation on the client-side
-    // via https://github.com/chriso/validator.js/blob/master/lib/isPostalCode.js#L54
-    // we have to double-escape the regex patterns here
-    // to output it as a string for the HTML pattern attribute
-    const POSTCODE_PATTERN = '(gir\\s?0aa|[a-zA-Z]{1,2}\\d[\\da-zA-Z]?\\s?(\\d[a-zA-Z]{2})?)';
-
-    return {
-        ...{
-            type: 'text',
-            attributes: {
-                size: 10,
-                autocomplete: 'postal-code',
-                pattern: POSTCODE_PATTERN
-            },
-            isRequired: true,
-            validator: validators.postcode
-        },
-        ...props
-    };
-}
-
-function addressFields(prefix) {
-    return [
-        {
-            name: `${prefix}-address-building-street`,
-            label: { en: 'Building and street', cy: '' },
-            type: 'text',
-            attributes: { size: 50 },
-            isRequired: true,
-            validator: validators.required({
-                en: 'Please provide a building and street',
-                cy: ''
-            })
-        },
-        {
-            name: `${prefix}-address-town-city`,
-            label: { en: 'Town or city', cy: '' },
-            type: 'text',
-            attributes: { size: 25 },
-            isRequired: true,
-            validator: validators.required({
-                en: 'Please provide a town or city',
-                cy: ''
-            })
-        },
-        {
-            name: `${prefix}-address-county`,
-            label: { en: 'County', cy: '' },
-            type: 'text',
-            attributes: { size: 25 },
-            isRequired: true,
-            validator: validators.required({
-                en: 'Please provide a county',
-                cy: ''
-            })
-        },
-        postcodeField({
-            name: `${prefix}-address-postcode`,
-            label: { en: 'Postcode', cy: '' }
-        })
-    ];
-}
-
-const FIELDS = {
-    projectStartDate: {
-        name: 'project-start-date',
-        label: {
-            en: 'When is the planned (or estimated) start date of your project?',
-            cy: '(WELSH) When is the planned (or estimated) start date of your project?'
-        },
-        explanation: {
-            en:
-                'This date needs to be at least 12 weeks from when you plan to submit your application. If your project is a one-off event, please tell us the date of the event.',
-            cy:
-                '(WELSH) This date needs to be at least 12 weeks from when you plan to submit your application. If your project is a one-off event, please tell us the date of the event.'
-        },
-        type: 'date',
-        attributes: {
-            min: moment()
-                .add(12, 'weeks')
-                .format('YYYY-MM-DD')
-        },
-        isRequired: true,
-        validator: validators.futureDate
-    },
-    projectPostcode: postcodeField({
-        name: 'project-postcode',
-        label: {
-            en: 'What is the postcode of the location where your project will take place?',
-            cy: '(WELSH) What is the postcode of the location where your project will take place?'
-        },
-        explanation: {
-            en:
-                'If your project will take place across different locations, please use the postcode where most of the project will take place.',
-            cy:
-                '(WELSH) If your project will take place across different locations, please use the postcode where most of the project will take place.'
-        }
-    }),
-    beneficiaryNumbers: {
-        name: 'beneficiary-numbers',
-        label: {
-            en: 'How many people will benefit from your project?',
-            cy: '(WELSH) How many people will benefit from your project?'
-        },
-        explanation: {
-            en: 'Please enter the exact figure, or the closest estimate.',
-            cy: '(WELSH) Please enter the exact figure, or the closest estimate.'
-        },
-        type: 'text',
-        isRequired: true,
-        validator: validators.required({
-            en: 'Please tell us how many people will benefit',
-            cy: ''
-        })
-    },
-    yourIdea: {
-        name: 'your-idea',
-        label: {
-            en: 'What would you like to do?',
-            cy: 'WELSH What would you like to do?'
-        },
-        explanation: {
-            en: `
-<p>When answering this question there are two key areas that we will use to make a decision on your project:</p>
-
-<p><strong>1. National Lottery Awards for all has three funding priorities and you must meet at least one of these. Please tell us how your project will:</strong></p>
-
-<ul>
-    <li>bring people together and build strong relationships in and across communities</li>
-    <li>improve the places and spaces that matter to communities</li>         
-    <li>enable more people to fulfil their potential by working to address issues at the earliest possible stage</li>     
-</ul>
-            
-<p><strong>2. It's important to us that you involve your community in the design, development and delivery of the activities you're planning, so please tell us how you've done this.</strong></p>
-
-<p>Here are some ideas about what else to tell us:</p>
-
-<ul>            
-    <li>How your project idea came about. Is it something new, or are you continuing something that has worked well previously?</li>
-    <li>If you are running a one-off event, what date it will take place</li>
-    <li>How long you expect your project to run</li>
-    <li>How you will make sure people know about your project and will attend</li>
-    <li>How you'll learn from your project and use this to shape future projects</li>
-</ul>
-            `,
-            cy: 'TODO'
-        },
-        type: 'textarea',
-        attributes: {
-            rows: 12
-        },
-        isRequired: true,
-        validator: validators.required({
-            en: 'Please tell us about your idea',
-            cy: ''
-        })
-    },
-    projectBudget: {
-        name: 'project-budget',
-        label: {
-            en: 'Please tell us the costs you would like us to fund',
-            cy: '(WELSH) Please tell us the costs you would like us to fund'
-        },
-        explanation: {
-            en: `
-<p>You should use budget headings, rather than a detailed list of items.</p>
-<p>For example, if you're applying for pens, pencils, paper and envelopes, using 'office supplies' is fine.</p>
-            `,
-            cy: 'TODO'
-        },
-        type: 'textarea',
-        attributes: {
-            rows: 12
-        },
-        isRequired: true,
-        validator: validators.required({
-            en: 'Field must be provided',
-            cy: ''
-        })
-    },
-    projectTotalCosts: {
-        name: 'project-total-costs',
-        label: {
-            en: 'Please tell us the total cost of your project.',
-            cy: '(WELSH) Please tell us the total cost of your project.'
-        },
-        explanation: {
-            en: `
-            <p>This is the cost of everything related to your project, even things you aren't asking us to fund.</p>
-
-            <p>For example, if you are asking us for £8,000 and you are getting £10,000 from another funder to cover additional costs, then your total project cost is £18,000. If you are asking us for £8,000 and there are no other costs then your total project cost is £8,000.</p>
-            `,
-            cy: 'TODO'
-        },
-        type: 'currency',
-        isRequired: true,
-        validator: validators.required({
-            en: 'Field must be provided',
-            cy: ''
-        })
-    },
-    organisationLegalName: {
-        name: 'organisation-legal-name',
-        label: {
-            en: 'What is the full legal name of your organisation?',
-            cy: '(WELSH) What is the full legal name of your organisation, as shown on your governing document?'
-        },
-        explanation: {
-            en: `
-            <p>This must be as shown on your <strong>governing document</strong>. Your governing document could be called one of several things, depending on the type of organisation you're applying on behalf of. It may be called a constitution, trust deed, memorandum and articles of association, or something else entirely.</p>
-            `,
-            cy: 'TODO'
-        },
-        type: 'text',
-        isRequired: true,
-        validator: validators.required({
-            en: 'Field must be provided',
-            cy: ''
-        })
-    },
-    organisationAlias: {
-        name: 'organisation-alias',
-        label: { en: 'Does your organisation use a different name in your day-to-day work?', cy: '' },
-        type: 'text',
-        isRequired: false,
-        validator: validators.optional
-    },
-    organisationType: {
-        name: 'organisation-type',
-        label: { en: 'What type of organisation are you?', cy: '(WELSH) What type of organisation are you?' },
-        type: 'radio',
-        options: values(ORGANISATION_TYPES),
-        validator: validators.required({
-            en: 'Field must be provided',
-            cy: ''
-        })
-    },
-    charityNumber: {
-        name: 'charity-number',
-        label: { en: 'Charity registration number', cy: '' },
-        explanation: {
-            en:
-                'If you are registered with OSCR, you only need to provide the last five digits of your registration number.',
-            cy: ''
-        },
-        type: 'number',
-        isRequired: true,
-        validator: validators.required({ en: 'Please provide a charity number', cy: '' })
-    },
-    companyNumber: {
-        name: 'company-number',
-        label: { en: 'Companies house number', cy: '' },
-        type: 'number',
-        isRequired: true,
-        validator: validators.required({ en: 'Please provide a company number', cy: '' })
-    },
-    accountingYearDate: {
-        name: 'accounting-year-date',
-        label: { en: 'What is your accounting year end date?', cy: '' },
-        type: 'date',
-        attributes: {
-            max: moment().format('YYYY-MM-DD')
-        },
-        isRequired: true,
-        validator: validators.pastDate
-    },
-    totalIncomeYear: {
-        name: 'total-income-year',
-        label: { en: 'What is your total income for the year?', cy: '' },
-        type: 'currency',
-        isRequired: true,
-        validator: validators.required({
-            en: 'Provide your total income for the year',
-            cy: ''
-        })
-    },
-    mainContactName: {
-        name: 'main-contact-name',
-        label: { en: 'Full name', cy: '' },
-        type: 'text',
-        attributes: {
-            autocomplete: 'name',
-            spellcheck: 'false'
-        },
-        isRequired: true,
-        validator: validators.required({
-            en: 'Enter your full name',
-            cy: ''
-        })
-    },
-    mainContactDob: {
-        name: 'main-contact-dob',
-        label: { en: 'Date of birth', cy: '' },
-        type: 'date',
-        attributes: {
-            max: moment()
-                .subtract(MIN_APPLICANT_AGE, 'years')
-                .format('YYYY-MM-DD')
-        },
-        isRequired: true,
-        validator: validators.dateOfBirth(MIN_APPLICANT_AGE)
-    },
-    mainContactEmail: {
-        name: 'main-contact-email',
-        label: { en: 'Email', cy: '' },
-        explanation: { en: 'We’ll use this whenever we get in touch about the project', cy: '' },
-        type: 'email',
-        isRequired: true,
-        validator: validators.emailRequired
-    },
-    mainContactPhone: {
-        name: 'main-contact-phone',
-        type: 'tel',
-        attributes: {
-            size: 30,
-            autocomplete: 'tel'
-        },
-        label: { en: 'UK telephone number', cy: '' },
-        isRequired: true,
-        validator: validators.required({
-            en: 'Please provide a phone number',
-            cy: ''
-        })
-    },
-    mainContactCommunicationNeeds: {
-        name: 'main-contact-communication-needs',
-        label: { en: 'Does this contact have any communication needs?', cy: '' },
-        type: 'radio',
-        options: [
-            { value: 'audiotape', label: { en: 'Audiotape', cy: '' } },
-            { value: 'braille', label: { en: 'Braille', cy: '' } },
-            { value: 'large-print', label: { en: 'Large print', cy: '' } }
-        ],
-        validator: validators.optional
-    },
-    legalContactName: {
-        name: 'legal-contact-name',
-        label: { en: 'Full name', cy: '' },
-        type: 'text',
-        attributes: {
-            autocomplete: 'name',
-            spellcheck: 'false'
-        },
-        isRequired: true,
-        validator: validators.required({
-            en: 'Enter your full name',
-            cy: ''
-        })
-    },
-    legalContactDob: {
-        name: 'legal-contact-dob',
-        label: { en: 'Date of birth', cy: '' },
-        type: 'date',
-        attributes: {
-            max: moment()
-                .subtract(MIN_APPLICANT_AGE, 'years')
-                .format('YYYY-MM-DD')
-        },
-        isRequired: true,
-        validator: validators.dateOfBirth(MIN_APPLICANT_AGE)
-    },
-    legalContactAddressBuildingStreet: {
-        name: 'legal-contact-address-building-street',
-        label: { en: 'Building and street', cy: '' },
-        type: 'text',
-        attributes: { size: 50 },
-        isRequired: true,
-        validator: validators.required({
-            en: 'Please provide a building and street',
-            cy: ''
-        })
-    },
-    legalContactAddressTownCity: {
-        name: 'legal-contact-address-town-city',
-        label: { en: 'Town or city', cy: '' },
-        type: 'text',
-        attributes: { size: 25 },
-        isRequired: true,
-        validator: validators.required({
-            en: 'Please provide a town or city',
-            cy: ''
-        })
-    },
-    legalContactAddressCounty: {
-        name: 'legal-contact-address-county',
-        label: { en: 'County', cy: '' },
-        type: 'text',
-        attributes: { size: 25 },
-        isRequired: true,
-        validator: validators.required({
-            en: 'Please provide a county',
-            cy: ''
-        })
-    },
-    legalContactAddressPostcode: postcodeField({
-        name: 'legal-contact-address-postcode',
-        label: { en: 'Postcode', cy: '' }
-    }),
-    legalContactEmail: {
-        name: 'legal-contact-email',
-        type: 'email',
-        label: { en: 'Email', cy: '' },
-        explanation: { en: 'We’ll use this whenever we get in touch about the project', cy: '' },
-        isRequired: true,
-        validator: validators.emailRequired
-    },
-    legalContactPhone: {
-        name: 'legal-contact-phone',
-        autocompleteName: 'tel',
-        type: 'tel',
-        size: 30,
-        label: { en: 'Contact number', cy: '' },
-        isRequired: true,
-        validator: validators.required({
-            en: 'Please provide a phone number',
-            cy: ''
-        })
-    },
-    legalContactCommunicationNeeds: {
-        name: 'legal-contact-communication-needs',
-        type: 'radio',
-        label: { en: 'Does this contact have any communication needs?', cy: '' },
-        options: [
-            { value: 'audiotape', label: { en: 'Audiotape', cy: '' } },
-            { value: 'braille', label: { en: 'Braille', cy: '' } },
-            { value: 'large-print', label: { en: 'Large print', cy: '' } }
-        ],
-        validator: validators.optional
-    },
-    bankAccountName: {
-        name: 'bank-account-name',
-        label: { en: 'Name on the bank account', cy: '' },
-        explanation: { en: 'Name of your organisation as it appears on your bank statement', cy: '' },
-        type: 'text',
-        isRequired: true,
-        validator: validators.required({
-            en: 'Please provide the name on the account',
-            cy: ''
-        })
-    },
-    bankSortCode: {
-        name: 'bank-sort-code',
-        label: { en: 'Sort code', cy: '' },
-        type: 'number',
-        attributes: {
-            size: 20
-        },
-        isRequired: true,
-        validator: validators.required({
-            en: 'Please provide a sort code',
-            cy: ''
-        })
-    },
-    bankAccountNumber: {
-        name: 'bank-account-number',
-        label: { en: 'Account number', cy: '' },
-        type: 'number',
-        isRequired: true,
-        validator: validators.required({
-            en: 'Please provide a bank account number',
-            cy: ''
-        })
-    },
-    bankBuildingSocietyNumber: {
-        name: 'bank-building-society-number',
-        label: { en: 'Building society number (if applicable)', cy: '' },
-        type: 'number',
-        explanation: {
-            en: 'This is only applicable if your organisation’s account is with a building society.',
-            cy: ''
-        },
-        validator: validators.optional
-    },
-    bankStatement: {
-        name: 'bank-statement',
-        label: { en: 'Upload a bank statement', cy: '' },
-        type: 'file',
-        isRequired: true,
-        validator: validators.required({
-            en: 'Please provide a bank statement',
-            cy: ''
-        })
-    },
-    terms: [
-        {
-            name: 'terms-agreement-1',
-            type: 'checkbox',
-            label: {
-                en:
-                    'You have been authorised by the governing body of your organisation (the board or committee that runs your organisation) to submit this application and to accept the Terms and Conditions set out above on their behalf.',
-                cy: ''
-            },
-            options: [{ value: 'yes', label: { en: 'I agree', cy: '' } }],
-            isRequired: true,
-            validator: validators.required({
-                en: 'Please provide your agreement',
-                cy: ''
-            })
-        },
-        {
-            name: 'terms-agreement-2',
-            type: 'checkbox',
-            label: {
-                en:
-                    'All the information you have provided in your application is accurate and complete; and you will notify us of any changes.',
-                cy: ''
-            },
-            options: [{ value: 'yes', label: { en: 'I agree', cy: '' } }],
-            isRequired: true,
-            validator: validators.required({
-                en: 'Please provide your agreement',
-                cy: ''
-            })
-        },
-        {
-            name: 'terms-agreement-3',
-            type: 'checkbox',
-            label: {
-                en:
-                    'You understand that we will use any personal information you have provided for the purposes described under the Data Protection Statement.',
-                cy: ''
-            },
-            options: [{ value: 'yes', label: { en: 'I agree', cy: '' } }],
-            isRequired: true,
-            validator: validators.required({
-                en: 'Please provide your agreement',
-                cy: ''
-            })
-        },
-        {
-            name: 'terms-agreement-4',
-            type: 'checkbox',
-            label: {
-                en:
-                    'If information about this application is requested under the Freedom of Information Act, we will release it in line with our Freedom of Information policy.',
-                cy: ''
-            },
-            options: [{ value: 'yes', label: { en: 'I agree', cy: '' } }],
-            isRequired: true,
-            validator: validators.required({
-                en: 'Please provide your agreement',
-                cy: ''
-            })
-        },
-        {
-            name: 'terms-person-name',
-            autocompleteName: 'name',
-            type: 'text',
-            label: { en: 'Full name of person completing this form', cy: '' },
-            isRequired: true,
-            validator: validators.required({
-                en: 'Enter the full name',
-                cy: ''
-            })
-        },
-        {
-            name: 'terms-person-position',
-            autocompleteName: 'position',
-            type: 'text',
-            label: { en: 'Position in organisation', cy: '' },
-            isRequired: true,
-            validator: validators.required({
-                en: 'Enter the position',
-                cy: ''
-            })
-        }
-    ],
-    applicationTitle: {
-        name: 'application-title',
-        type: 'text',
-        label: { en: 'Funding application title', cy: '' },
-        isRequired: true,
-        validator: validators.required({
-            en: 'Please provide a title',
-            cy: ''
-        })
-    }
-};
 
 /**
  * @type Section
@@ -685,7 +68,7 @@ const sectionProject = {
                     /**
                      * @type Array<Field>
                      */
-                    fields: [FIELDS.projectStartDate, FIELDS.projectPostcode]
+                    fields: [allFields.projectStartDate, allFields.projectPostcode]
                 }
             ]
         },
@@ -694,7 +77,7 @@ const sectionProject = {
             fieldsets: [
                 {
                     legend: { en: 'Your idea', cy: '(WELSH) Your idea' },
-                    fields: [FIELDS.yourIdea]
+                    fields: [allFields.yourIdea]
                 }
             ]
         },
@@ -703,7 +86,7 @@ const sectionProject = {
             fieldsets: [
                 {
                     legend: { en: 'Project costs', cy: '(WELSH) Project costs' },
-                    fields: [FIELDS.projectBudget, FIELDS.projectTotalCosts]
+                    fields: [allFields.projectBudget, allFields.projectTotalCosts]
                 }
             ]
         }
@@ -727,7 +110,7 @@ const sectionOrganisation = {
             fieldsets: [
                 {
                     legend: { en: 'Organisation details', cy: '' },
-                    fields: [FIELDS.organisationLegalName, FIELDS.organisationAlias]
+                    fields: [allFields.organisationLegalName, allFields.organisationAlias]
                 },
                 {
                     legend: { en: 'What is the main or registered address of your organisation?', cy: '' },
@@ -740,19 +123,19 @@ const sectionOrganisation = {
             fieldsets: [
                 {
                     legend: { en: 'Organisation type', cy: '' },
-                    fields: [FIELDS.organisationType]
+                    fields: [allFields.organisationType]
                 }
             ]
         },
         {
             title: { en: 'Charity number', cy: '' },
             matchesCondition: function(formData) {
-                return formData['organisation-type'] === ORGANISATION_TYPES.unincorporatedRegisteredCharity.value;
+                return formData['organisation-type'] === organisationTypes.unincorporatedRegisteredCharity.value;
             },
             fieldsets: [
                 {
                     legend: { en: 'Charity number', cy: '' },
-                    fields: [FIELDS.charityNumber]
+                    fields: [allFields.charityNumber]
                 }
             ]
         },
@@ -761,9 +144,9 @@ const sectionOrganisation = {
             matchesCondition: function(formData) {
                 return includes(
                     [
-                        ORGANISATION_TYPES.charitableIncorporatedOrganisation.value,
-                        ORGANISATION_TYPES.notForProfitCompany.value,
-                        ORGANISATION_TYPES.communityInterestCompany.value
+                        organisationTypes.charitableIncorporatedOrganisation.value,
+                        organisationTypes.notForProfitCompany.value,
+                        organisationTypes.communityInterestCompany.value
                     ],
                     formData['organisation-type']
                 );
@@ -771,7 +154,7 @@ const sectionOrganisation = {
             fieldsets: [
                 {
                     legend: { en: 'Company number', cy: '' },
-                    fields: [FIELDS.companyNumber]
+                    fields: [allFields.companyNumber]
                 }
             ]
         },
@@ -780,7 +163,7 @@ const sectionOrganisation = {
             fieldsets: [
                 {
                     legend: { en: 'Organisation finances', cy: '' },
-                    fields: [FIELDS.accountingYearDate, FIELDS.totalIncomeYear]
+                    fields: [allFields.accountingYearDate, allFields.totalIncomeYear]
                 }
             ]
         }
@@ -811,7 +194,7 @@ const sectionMainContact = {
 <p>The main contact must be unconnected to the legally responsible contact. By ‘unconnected’ we mean not related by blood, marriage, in a long-term relationship or people living together at the same address.</p>`,
                         cy: ''
                     },
-                    fields: [FIELDS.mainContactName, FIELDS.mainContactDob]
+                    fields: [allFields.mainContactName, allFields.mainContactDob]
                 },
                 {
                     legend: { en: 'Address', cy: '' },
@@ -819,7 +202,11 @@ const sectionMainContact = {
                 },
                 {
                     legend: { en: 'Contact details', cy: '' },
-                    fields: [FIELDS.mainContactEmail, FIELDS.mainContactPhone, FIELDS.mainContactCommunicationNeeds]
+                    fields: [
+                        allFields.mainContactEmail,
+                        allFields.mainContactPhone,
+                        allFields.mainContactCommunicationNeeds
+                    ]
                 }
             ]
         }
@@ -852,7 +239,7 @@ const sectionLegalContact = {
 <p>The position held by the legally responsible contact is dependent on the type of organisation you are applying on behalf of. The options given to you for selection are based on this.</p>`,
                         cy: ''
                     },
-                    fields: [FIELDS.legalContactName, FIELDS.legalContactDob]
+                    fields: [allFields.legalContactName, allFields.legalContactDob]
                 },
                 {
                     legend: { en: 'Address', cy: '' },
@@ -860,7 +247,11 @@ const sectionLegalContact = {
                 },
                 {
                     legend: { en: 'Contact details', cy: '' },
-                    fields: [FIELDS.legalContactEmail, FIELDS.legalContactPhone, FIELDS.legalContactCommunicationNeeds]
+                    fields: [
+                        allFields.legalContactEmail,
+                        allFields.legalContactPhone,
+                        allFields.legalContactCommunicationNeeds
+                    ]
                 }
             ]
         }
@@ -890,10 +281,10 @@ const sectionBankDetails = {
                         cy: ''
                     },
                     fields: [
-                        FIELDS.bankAccountName,
-                        FIELDS.bankSortCode,
-                        FIELDS.bankAccountNumber,
-                        FIELDS.bankBuildingSocietyNumber
+                        allFields.bankAccountName,
+                        allFields.bankSortCode,
+                        allFields.bankAccountNumber,
+                        allFields.bankBuildingSocietyNumber
                     ]
                 }
             ]
@@ -924,38 +315,113 @@ const sectionBankDetails = {
                         `,
                         cy: ''
                     },
-                    fields: [FIELDS.bankStatement]
+                    fields: [allFields.bankStatement]
                 }
             ]
         }
     ]
 };
 
-/**
- * @typedef {Object} FormModel
- * @property {String} sessionKey
- * @property {LocaleString} title
- * @property {boolean} isBilingual
- * @property {Array<Section>} sections
- * @property {Array<Field>} termsFields
- * @property {Field} titleField
- * @property {Object} schema
- * @property {Function} processor
- * @property {Object} startPage
- * @property {Object} successStep
- */
-
 module.exports = {
-    /**
-     * @type FormModel
-     */
     formModel: {
         sessionKey: SESSION_KEY,
         title: { en: 'National Lottery Awards for All', cy: '' },
         isBilingual: true,
         sections: [sectionProject, sectionOrganisation, sectionMainContact, sectionLegalContact, sectionBankDetails],
-        termsFields: FIELDS.terms,
-        titleField: FIELDS.applicationTitle,
+        termsFields: [
+            {
+                name: 'terms-agreement-1',
+                type: 'checkbox',
+                label: {
+                    en:
+                        'You have been authorised by the governing body of your organisation (the board or committee that runs your organisation) to submit this application and to accept the Terms and Conditions set out above on their behalf.',
+                    cy: ''
+                },
+                options: [{ value: 'yes', label: { en: 'I agree', cy: '' } }],
+                isRequired: true,
+                validator: validatorsLegacy.required({
+                    en: 'Please provide your agreement',
+                    cy: ''
+                })
+            },
+            {
+                name: 'terms-agreement-2',
+                type: 'checkbox',
+                label: {
+                    en:
+                        'All the information you have provided in your application is accurate and complete; and you will notify us of any changes.',
+                    cy: ''
+                },
+                options: [{ value: 'yes', label: { en: 'I agree', cy: '' } }],
+                isRequired: true,
+                validator: validatorsLegacy.required({
+                    en: 'Please provide your agreement',
+                    cy: ''
+                })
+            },
+            {
+                name: 'terms-agreement-3',
+                type: 'checkbox',
+                label: {
+                    en:
+                        'You understand that we will use any personal information you have provided for the purposes described under the Data Protection Statement.',
+                    cy: ''
+                },
+                options: [{ value: 'yes', label: { en: 'I agree', cy: '' } }],
+                isRequired: true,
+                validator: validatorsLegacy.required({
+                    en: 'Please provide your agreement',
+                    cy: ''
+                })
+            },
+            {
+                name: 'terms-agreement-4',
+                type: 'checkbox',
+                label: {
+                    en:
+                        'If information about this application is requested under the Freedom of Information Act, we will release it in line with our Freedom of Information policy.',
+                    cy: ''
+                },
+                options: [{ value: 'yes', label: { en: 'I agree', cy: '' } }],
+                isRequired: true,
+                validator: validatorsLegacy.required({
+                    en: 'Please provide your agreement',
+                    cy: ''
+                })
+            },
+            {
+                name: 'terms-person-name',
+                autocompleteName: 'name',
+                type: 'text',
+                label: { en: 'Full name of person completing this form', cy: '' },
+                isRequired: true,
+                validator: validatorsLegacy.required({
+                    en: 'Enter the full name',
+                    cy: ''
+                })
+            },
+            {
+                name: 'terms-person-position',
+                autocompleteName: 'position',
+                type: 'text',
+                label: { en: 'Position in organisation', cy: '' },
+                isRequired: true,
+                validator: validatorsLegacy.required({
+                    en: 'Enter the position',
+                    cy: ''
+                })
+            }
+        ],
+        titleField: {
+            name: 'application-title',
+            type: 'text',
+            label: { en: 'Funding application title', cy: '' },
+            isRequired: true,
+            validator: validatorsLegacy.required({
+                en: 'Please provide a title',
+                cy: ''
+            })
+        },
         schema: schema,
         processor: processor,
         startPage: { template: path.resolve(__dirname, '../views/startpage') },
